pyproject.toml
[project]
name = "workflow-certif"
version = "0.1.0"
description = "Pipeline de certification industrialisé pour CAF"
authors = [{name = "Florian Maillard"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["pandas", "openpyxl", "pyyaml"]

[project.scripts]
run-certif = "main:main"

[tool.setuptools]
packages = ["scripts", "core", "core.etapes"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
